airflow+postgres+bq/dags/ebird_analytics_dwh/requirements.txt
pandas
findspark
pyspark
requests
apache-airflow-providers-postgres
google-cloud-bigquery
google-cloud-storage
